+++ index.html
@@ -227,7 +227,7 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-  <script src="js/main.js" defer></script>
+  <script src="js/main.js" type="module"></script>
   <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.20/lodash.min.js"></script>
 </body>
 </html>
