+++ js/data.js
@@ -1,0 +1,52 @@
+const TITLES = [
+  'Объект1',
+  'Объект2',
+  'Объект3',
+  'Объект4',
+  'Объект5',
+  'Объект6',
+  'Объект7',
+  'Объект8',
+  'Объект9',
+  'Объект10',
+];
+
+const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow'];
+
+const HOURS = [
+  '12:00',
+  '13:00',
+  '14:00'];
+
+const FEATURE = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner'];
+
+const DESCRIPTIONS = [
+  'описание помещения1',
+  'описание помещения2',
+  'описание помещения3',
+  'описание помещения4',
+  'описание помещения5',
+  'описание помещения6',
+  'описание помещения7',
+  'описание помещения8',
+  'описание помещения9',
+  'описание помещения10'];
+
+const PHOTO = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg'];
+
+const SIMILAR_OFFERS_COUNT = 10;
+
+export {TITLES, TYPES, HOURS, FEATURE, DESCRIPTIONS, PHOTO, SIMILAR_OFFERS_COUNT}
